Thang

+++ app/src/main/java/com/oceantech/tracking/ui/tracking/TrackingFragment.kt
@@ -1,22 +1,19 @@
 package com.oceantech.tracking.ui.tracking
 
+import android.app.AlertDialog
+import android.content.DialogInterface
+import android.location.GnssAntennaInfo.Listener
 import android.os.Bundle
 import android.util.Log
 import android.view.Gravity
 import android.view.LayoutInflater
-import android.view.Menu
-import android.view.MenuInflater
-import android.view.MenuItem
 import android.view.View
 import android.view.ViewGroup
 import android.widget.LinearLayout
 import android.widget.PopupWindow
 import android.widget.Toast
-import androidx.annotation.MenuRes
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.content.ContextCompat.getDrawable
-import androidx.core.content.ContextCompat.getSystemService
-import androidx.lifecycle.lifecycleScope
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.airbnb.mvrx.Fail
 import com.airbnb.mvrx.Loading
@@ -27,22 +24,18 @@ import com.google.android.material.textview.MaterialTextView
 import com.oceantech.tracking.R
 import com.oceantech.tracking.core.TrackingBaseFragment
 import com.oceantech.tracking.data.model.Tracking
-import com.oceantech.tracking.databinding.FragmentAllTrackingBinding
+import com.oceantech.tracking.databinding.FragmentTrackingBinding
 import com.oceantech.tracking.ui.home.HomeViewAction
-import com.oceantech.tracking.ui.home.HomeViewEvent
 import com.oceantech.tracking.ui.home.HomeViewModel
-import kotlinx.coroutines.flow.collectLatest
 
-class AllTrackingFragment : TrackingBaseFragment<FragmentAllTrackingBinding>() {
+class TrackingFragment : TrackingBaseFragment<FragmentTrackingBinding>() {
     private val viewModel:HomeViewModel by activityViewModel()
     private lateinit var trackingAdapter:TrackingAdapter
     private lateinit var trackings:List<Tracking>
-
     override fun getBinding(
         inflater: LayoutInflater,
         container: ViewGroup?
-    ): FragmentAllTrackingBinding = FragmentAllTrackingBinding.inflate(inflater, container, false)
-
+    ): FragmentTrackingBinding = FragmentTrackingBinding.inflate(inflater, container, false)
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
         trackings = listOf()
@@ -51,6 +44,12 @@ class AllTrackingFragment : TrackingBaseFragment<FragmentAllTrackingBinding>() {
             layoutManager = LinearLayoutManager(requireContext())
             adapter = trackingAdapter
         }
+
+        viewModel.trackings.observe(viewLifecycleOwner){
+            trackingAdapter = TrackingAdapter(it,requireContext(), showMenu)
+            views.recyclerView.adapter = trackingAdapter
+        }
+        viewModel.handleAllTracking()
     }
 
     private val showMenu:(View,Tracking) -> Unit = {view,tracking ->
@@ -71,14 +70,20 @@ class AllTrackingFragment : TrackingBaseFragment<FragmentAllTrackingBinding>() {
         popup.showAsDropDown(v, 5, -5, Gravity.CENTER_HORIZONTAL)
 
         view.findViewById<MaterialTextView>(R.id.to_track_update).setOnClickListener {
-
+            viewModel.handleReturnUpdate(tracking.content!!, tracking.id!!)
         }
 
         view.findViewById<MaterialTextView>(R.id.to_track_delete).setOnClickListener {
-            deleteTracking(tracking)
+            AlertDialog.Builder(requireContext())
+                .setTitle(R.string.confirm)
+                .setMessage(R.string.confirm_delete)
+                .setNegativeButton(R.string.no, null)
+                .setPositiveButton(R.string.yes, DialogInterface.OnClickListener { dialog, which ->
+                    deleteTracking(tracking)
+                })
+                .show()
         }
     }
-
     private fun deleteTracking(tracking: Tracking) {
         viewModel.handle(HomeViewAction.DeleteTracking(tracking.id!!))
     }
@@ -86,11 +91,6 @@ class AllTrackingFragment : TrackingBaseFragment<FragmentAllTrackingBinding>() {
     override fun invalidate():Unit = withState(viewModel){
         when(it.allTracking){
             is Success -> {
-                it.allTracking.invoke().let { it ->
-                    trackings = it
-                    trackingAdapter = TrackingAdapter(trackings,requireContext(), showMenu)
-                    views.recyclerView.adapter = trackingAdapter
-                }
                 dismissLoadingDialog()
             }
             is Loading -> {
@@ -105,14 +105,14 @@ class AllTrackingFragment : TrackingBaseFragment<FragmentAllTrackingBinding>() {
                 it.asyncDeleteTracking.invoke().let {
                     Toast.makeText(requireContext(), getString(R.string.delete_success), Toast.LENGTH_SHORT).show()
                     Log.i("state of deleted: ", "success delete tracking id ${it.id}")
-                    dismissLoadingDialog()
                 }
+                dismissLoadingDialog()
             }
             is Loading -> {
+                viewModel.handleAllTracking()
                 showLoadingDialog()
             }
             is Fail -> {
-                Log.i("state of deleted: ", "fail ")
                 dismissLoadingDialog()
             }
         }
